Add descriptions to recipes
Without any context to what a recipe is and what it is made of, there
won't be any interest gauged.

By adding a Description section for each recipe, this solves that
issue.

P.S: I forgot to mention in the commit prior that I added images to
each recipe so a viewer can see what a recipe looks like.